backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
asyncpg==0.31.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
passlib[bcrypt]==1.7.4
pydantic[email]==2.12.5
python-dotenv==1.2.2
google-generativeai==0.8.6
httpx==0.28.1
